refactor(workspace): extract bounded cache microcrate
