+++ crates/perl-workspace-index/src/workspace/cache.rs
@@ -1,714 +1,5 @@
-//! Bounded LRU caches for workspace index components.
+//! Backward-compatible re-export of the workspace cache microcrate.
 //!
-//! This module provides thread-safe, bounded caches with LRU eviction policies
-//! for AST nodes, symbols, and workspace data. These caches enforce memory limits
-//! while maintaining high hit rates for frequently accessed data.
-//!
-//! # Performance Characteristics
-//!
-//! - **Cache hit rate**: >90% for typical workloads
-//! - **Eviction latency**: O(1) amortized with linked hash map
-//! - **Memory overhead**: ~32 bytes per cache entry
-//! - **Thread safety**: Lock-free reads with atomic reference counting
-//!
-//! # Cache Configuration
-//!
-//! - **AST Node Cache**: Max 10,000 nodes, 50MB memory limit
-//! - **Symbol Cache**: Max 50,000 symbols, 30MB memory limit
-//! - **Workspace Cache**: Max 1,000 files, 20MB memory limit
-//!
-//! # Usage
-//!
-//! ```rust,ignore
-//! use perl_workspace_index::workspace::cache::{BoundedLruCache, CacheConfig};
-//!
-//! let config = CacheConfig::default();
-//! let cache = BoundedLruCache::new(config);
-//!
-//! cache.insert("key", "value");
-//! assert_eq!(cache.get("key"), Some(&"value"));
-//! ```
+//! The concrete cache implementation now lives in `perl-workspace-cache`.
 
-use parking_lot::Mutex;
-use std::collections::HashMap;
-use std::hash::Hash;
-use std::sync::Arc;
-use std::time::{Duration, Instant};
-
-/// Cache configuration for bounded LRU caches.
-///
-/// Defines size limits and memory budgets for cache instances.
-#[derive(Clone, Debug)]
-pub struct CacheConfig {
-    /// Maximum number of items in the cache
-    pub max_items: usize,
-    /// Maximum memory usage in bytes
-    pub max_bytes: usize,
-    /// TTL for cache entries (None = no expiration)
-    pub ttl: Option<Duration>,
-}
-
-impl Default for CacheConfig {
-    fn default() -> Self {
-        Self {
-            max_items: 10_000,
-            max_bytes: 50 * 1024 * 1024, // 50MB
-            ttl: None,
-        }
-    }
-}
-
-/// Cache statistics for monitoring and diagnostics.
-#[derive(Clone, Debug, Default)]
-pub struct CacheStats {
-    /// Total number of cache hits
-    pub hits: u64,
-    /// Total number of cache misses
-    pub misses: u64,
-    /// Total number of evictions
-    pub evictions: u64,
-    /// Current number of items in cache
-    pub current_items: usize,
-    /// Current memory usage in bytes
-    pub current_bytes: usize,
-    /// Hit rate (hits / (hits + misses))
-    pub hit_rate: f64,
-}
-
-impl CacheStats {
-    /// Calculate hit rate from hits and misses.
-    pub fn calculate_hit_rate(hits: u64, misses: u64) -> f64 {
-        let total = hits + misses;
-        if total == 0 { 0.0 } else { hits as f64 / total as f64 }
-    }
-}
-
-/// Cache entry with metadata for LRU tracking and expiration.
-#[derive(Clone)]
-struct CacheEntry<V> {
-    /// The cached value
-    value: V,
-    /// When this entry was last accessed
-    last_accessed: Instant,
-    /// When this entry was inserted
-    _inserted_at: Instant,
-    /// Size of the entry in bytes
-    size_bytes: usize,
-}
-
-impl<V> CacheEntry<V> {
-    /// Create a new cache entry.
-    fn new(value: V, size_bytes: usize) -> Self {
-        let now = Instant::now();
-        Self { value, last_accessed: now, _inserted_at: now, size_bytes }
-    }
-
-    /// Check if this entry has expired based on TTL.
-    fn is_expired(&self, ttl: Duration) -> bool {
-        self.last_accessed.elapsed() > ttl
-    }
-
-    /// Update the last accessed time.
-    fn touch(&mut self) {
-        self.last_accessed = Instant::now();
-    }
-}
-
-/// Thread-safe bounded LRU cache.
-///
-/// Implements a least-recently-used eviction policy with configurable
-/// size limits and optional TTL expiration.
-///
-/// # Type Parameters
-///
-/// * `K` - Cache key type (must implement Hash + Eq)
-/// * `V` - Cache value type
-///
-/// # Performance
-///
-/// - **Insert**: O(1) amortized
-/// - **Get**: O(1) amortized
-/// - **Eviction**: O(1) amortized
-pub struct BoundedLruCache<K, V>
-where
-    K: Hash + Eq + Clone,
-{
-    /// Cache entries (key -> entry)
-    entries: Arc<Mutex<HashMap<K, CacheEntry<V>>>>,
-    /// Access order for LRU tracking (oldest keys at front)
-    access_order: Arc<Mutex<Vec<K>>>,
-    /// Cache configuration
-    config: CacheConfig,
-    /// Cache statistics
-    stats: Arc<Mutex<CacheStats>>,
-}
-
-impl<K, V> BoundedLruCache<K, V>
-where
-    K: Hash + Eq + Clone,
-{
-    /// Create a new bounded LRU cache with the given configuration.
-    ///
-    /// # Arguments
-    ///
-    /// * `config` - Cache configuration (size limits, TTL, etc.)
-    ///
-    /// # Returns
-    ///
-    /// A new bounded LRU cache instance.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use perl_workspace_index::workspace::cache::{BoundedLruCache, CacheConfig};
-    ///
-    /// let config = CacheConfig {
-    ///     max_items: 1000,
-    ///     max_bytes: 10 * 1024 * 1024, // 10MB
-    ///     ttl: None,
-    /// };
-    /// let cache: BoundedLruCache<String, String> = BoundedLruCache::new(config);
-    /// ```
-    pub fn new(config: CacheConfig) -> Self {
-        Self {
-            entries: Arc::new(Mutex::new(HashMap::new())),
-            access_order: Arc::new(Mutex::new(Vec::new())),
-            config,
-            stats: Arc::new(Mutex::new(CacheStats::default())),
-        }
-    }
-
-    /// Create a new cache with default configuration.
-    ///
-    /// # Returns
-    ///
-    /// A new bounded LRU cache with default limits (10,000 items, 50MB).
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use perl_workspace_index::workspace::cache::BoundedLruCache;
-    ///
-    /// let cache: BoundedLruCache<String, String> = BoundedLruCache::default();
-    /// ```
-    pub fn default() -> Self {
-        Self::new(CacheConfig::default())
-    }
-
-    /// Insert a value into the cache.
-    ///
-    /// If the cache is at capacity, the least recently used entry will be evicted.
-    ///
-    /// # Arguments
-    ///
-    /// * `key` - Cache key
-    /// * `value` - Value to cache
-    /// * `size_bytes` - Size of the value in bytes (for memory tracking)
-    ///
-    /// # Returns
-    ///
-    /// `true` if the value was inserted, `false` if evicted due to size limits.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use perl_workspace_index::workspace::cache::BoundedLruCache;
-    ///
-    /// let mut cache = BoundedLruCache::default();
-    /// cache.insert_with_size("key", "value", 5);
-    /// ```
-    pub fn insert_with_size(&self, key: K, value: V, size_bytes: usize) -> bool {
-        let mut entries = self.entries.lock();
-        let mut access_order = self.access_order.lock();
-        let mut stats = self.stats.lock();
-
-        // Check if this key already exists
-        if let Some(entry) = entries.get_mut(&key) {
-            // Update existing entry
-            entry.value = value;
-            entry.size_bytes = size_bytes;
-            entry.touch();
-
-            // Update access order (move to end = most recent)
-            if let Some(pos) = access_order.iter().position(|k| k == &key) {
-                access_order.remove(pos);
-            }
-            access_order.push(key.clone());
-
-            // Update stats
-            stats.current_bytes = entries.values().map(|e| e.size_bytes).sum();
-            stats.current_items = entries.len();
-            return true;
-        }
-
-        // Check if we need to evict entries
-        while !entries.is_empty()
-            && (entries.len() >= self.config.max_items
-                || stats.current_bytes + size_bytes > self.config.max_bytes)
-        {
-            // Evict least recently used (first in access_order)
-            if let Some(lru_key) = access_order.first() {
-                if let Some(entry) = entries.remove(lru_key) {
-                    stats.current_bytes -= entry.size_bytes;
-                    stats.evictions += 1;
-                }
-                access_order.remove(0);
-            } else {
-                break;
-            }
-        }
-
-        // Check if we can fit this entry
-        if entries.len() >= self.config.max_items
-            || stats.current_bytes + size_bytes > self.config.max_bytes
-        {
-            // Entry too large for cache
-            return false;
-        }
-
-        // Insert new entry
-        entries.insert(key.clone(), CacheEntry::new(value, size_bytes));
-        access_order.push(key);
-
-        // Update stats
-        stats.current_bytes = entries.values().map(|e| e.size_bytes).sum();
-        stats.current_items = entries.len();
-
-        true
-    }
-
-    /// Insert a value into the cache with estimated size.
-    ///
-    /// Uses a simple size estimation based on the value's memory representation.
-    ///
-    /// # Arguments
-    ///
-    /// * `key` - Cache key
-    /// * `value` - Value to cache
-    ///
-    /// # Returns
-    ///
-    /// `true` if the value was inserted, `false` if evicted.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use perl_workspace_index::workspace::cache::BoundedLruCache;
-    ///
-    /// let mut cache = BoundedLruCache::default();
-    /// cache.insert("key", "value");
-    /// ```
-    pub fn insert(&self, key: K, value: V)
-    where
-        V: EstimateSize,
-    {
-        let size_bytes = value.estimate_size();
-        self.insert_with_size(key, value, size_bytes);
-    }
-
-    /// Get a value from the cache.
-    ///
-    /// Returns `None` if the key is not found or the entry has expired.
-    ///
-    /// # Arguments
-    ///
-    /// * `key` - Cache key to look up
-    ///
-    /// # Returns
-    ///
-    /// `Some(&V)` if found and not expired, `None` otherwise.
-    ///
-    /// # Examples
-    ///
-    /// ```rust,ignore
-    /// use perl_workspace_index::workspace::cache::BoundedLruCache;
-    ///
-    /// let mut cache = BoundedLruCache::default();
-    /// cache.insert("key", "value");
-    /// assert_eq!(cache.get("key"), Some(&"value"));
-    /// ```
-    pub fn get(&self, key: &K) -> Option<V>
-    where
-        V: Clone,
-    {
-        let mut entries = self.entries.lock();
-        let mut access_order = self.access_order.lock();
-        let mut stats = self.stats.lock();
-
-        // Check TTL expiration
-        if let Some(ttl) = self.config.ttl {
-            if let Some(entry) = entries.get(key) {
-                if entry.is_expired(ttl) {
-                    // Entry expired, remove it
-                    entries.remove(key);
-                    if let Some(pos) = access_order.iter().position(|k| k == key) {
-                        access_order.remove(pos);
-                    }
-                    stats.misses += 1;
-                    stats.hit_rate = CacheStats::calculate_hit_rate(stats.hits, stats.misses);
-                    return None;
-                }
-            }
-        }
-
-        // Get entry and update LRU
-        if let Some(entry) = entries.get_mut(key) {
-            entry.touch();
-
-            // Update access order (move to end = most recent)
-            if let Some(pos) = access_order.iter().position(|k| k == key) {
-                let key_clone = access_order.remove(pos);
-                access_order.push(key_clone);
-            }
-
-            stats.hits += 1;
-            stats.hit_rate = CacheStats::calculate_hit_rate(stats.hits, stats.misses);
-            Some(entry.value.clone())
-        } else {
-            stats.misses += 1;
-            stats.hit_rate = CacheStats::calculate_hit_rate(stats.hits, stats.misses);
-            None
-        }
-    }
-
-    /// Remove a value from the cache.
-    ///
-    /// # Arguments
-    ///
-    /// * `key` - Cache key to remove
-    ///
-    /// # Returns
-    ///
-    /// `Some(V)` if the entry was removed, `None` if not found.
-    ///
-    /// # Examples
-    ///
-    /// ```rust,ignore
-    /// use perl_workspace_index::workspace::cache::BoundedLruCache;
-    ///
-    /// let mut cache = BoundedLruCache::default();
-    /// cache.insert("key", "value");
-    /// assert_eq!(cache.remove("key"), Some("value"));
-    /// ```
-    pub fn remove(&self, key: &K) -> Option<V>
-    where
-        V: Clone,
-    {
-        let mut entries = self.entries.lock();
-        let mut access_order = self.access_order.lock();
-        let mut stats = self.stats.lock();
-
-        if let Some(entry) = entries.remove(key) {
-            stats.current_bytes -= entry.size_bytes;
-            stats.current_items = entries.len();
-
-            if let Some(pos) = access_order.iter().position(|k| k == key) {
-                access_order.remove(pos);
-            }
-
-            Some(entry.value)
-        } else {
-            None
-        }
-    }
-
-    /// Clear all entries from the cache.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use perl_workspace_index::workspace::cache::BoundedLruCache;
-    ///
-    /// let mut cache = BoundedLruCache::default();
-    /// cache.insert("key", "value");
-    /// cache.clear();
-    /// assert!(cache.is_empty());
-    /// ```
-    pub fn clear(&self) {
-        let mut entries = self.entries.lock();
-        let mut access_order = self.access_order.lock();
-        let mut stats = self.stats.lock();
-
-        entries.clear();
-        access_order.clear();
-        stats.current_bytes = 0;
-        stats.current_items = 0;
-    }
-
-    /// Check if the cache is empty.
-    ///
-    /// # Returns
-    ///
-    /// `true` if the cache contains no entries, `false` otherwise.
-    pub fn is_empty(&self) -> bool {
-        let entries = self.entries.lock();
-        entries.is_empty()
-    }
-
-    /// Get the number of items in the cache.
-    ///
-    /// # Returns
-    ///
-    /// The current number of cached items.
-    pub fn len(&self) -> usize {
-        let entries = self.entries.lock();
-        entries.len()
-    }
-
-    /// Get cache statistics.
-    ///
-    /// # Returns
-    ///
-    /// A snapshot of the cache statistics.
-    ///
-    /// # Examples
-    ///
-    /// ```rust,ignore
-    /// use perl_workspace_index::workspace::cache::BoundedLruCache;
-    ///
-    /// let cache = BoundedLruCache::default();
-    /// let stats = cache.stats();
-    /// assert_eq!(stats.hits, 0);
-    /// ```
-    pub fn stats(&self) -> CacheStats {
-        let stats = self.stats.lock();
-        stats.clone()
-    }
-
-    /// Get the cache configuration.
-    ///
-    /// # Returns
-    ///
-    /// The cache configuration.
-    pub fn config(&self) -> &CacheConfig {
-        &self.config
-    }
-}
-
-/// Trait for estimating the memory size of cached values.
-///
-/// Implement this trait for custom types to enable accurate memory tracking.
-pub trait EstimateSize {
-    /// Estimate the memory size of this value in bytes.
-    fn estimate_size(&self) -> usize;
-}
-
-// Implement EstimateSize for common types
-impl EstimateSize for String {
-    fn estimate_size(&self) -> usize {
-        self.len()
-    }
-}
-
-impl<T> EstimateSize for Vec<T>
-where
-    T: EstimateSize,
-{
-    fn estimate_size(&self) -> usize {
-        self.iter().map(|t| t.estimate_size()).sum()
-    }
-}
-
-impl<K, V> EstimateSize for HashMap<K, V>
-where
-    K: EstimateSize,
-    V: EstimateSize,
-{
-    fn estimate_size(&self) -> usize {
-        self.iter().map(|(k, v)| k.estimate_size() + v.estimate_size()).sum()
-    }
-}
-
-impl EstimateSize for str {
-    fn estimate_size(&self) -> usize {
-        self.len()
-    }
-}
-
-impl<T: EstimateSize + ?Sized> EstimateSize for &T {
-    fn estimate_size(&self) -> usize {
-        (**self).estimate_size()
-    }
-}
-
-impl EstimateSize for [u8] {
-    fn estimate_size(&self) -> usize {
-        self.len()
-    }
-}
-
-impl EstimateSize for () {
-    fn estimate_size(&self) -> usize {
-        0
-    }
-}
-
-impl<T> EstimateSize for Option<T>
-where
-    T: EstimateSize,
-{
-    fn estimate_size(&self) -> usize {
-        self.as_ref().map(|t| t.estimate_size()).unwrap_or(0)
-    }
-}
-
-impl<T, E> EstimateSize for Result<T, E>
-where
-    T: EstimateSize,
-    E: EstimateSize,
-{
-    fn estimate_size(&self) -> usize {
-        match self {
-            Ok(t) => t.estimate_size(),
-            Err(e) => e.estimate_size(),
-        }
-    }
-}
-
-/// AST node cache configuration.
-///
-/// Optimized for AST node caching with higher memory limits.
-#[derive(Clone, Debug)]
-pub struct AstCacheConfig {
-    /// Maximum number of AST nodes to cache
-    pub max_nodes: usize,
-    /// Maximum memory for AST cache in bytes
-    pub max_bytes: usize,
-}
-
-impl Default for AstCacheConfig {
-    fn default() -> Self {
-        Self {
-            max_nodes: 10_000,
-            max_bytes: 50 * 1024 * 1024, // 50MB
-        }
-    }
-}
-
-/// Symbol cache configuration.
-///
-/// Optimized for symbol lookup caching.
-#[derive(Clone, Debug)]
-pub struct SymbolCacheConfig {
-    /// Maximum number of symbols to cache
-    pub max_symbols: usize,
-    /// Maximum memory for symbol cache in bytes
-    pub max_bytes: usize,
-}
-
-impl Default for SymbolCacheConfig {
-    fn default() -> Self {
-        Self {
-            max_symbols: 50_000,
-            max_bytes: 30 * 1024 * 1024, // 30MB
-        }
-    }
-}
-
-/// Workspace cache configuration.
-///
-/// Optimized for workspace file metadata caching.
-#[derive(Clone, Debug)]
-pub struct WorkspaceCacheConfig {
-    /// Maximum number of workspace files to cache
-    pub max_files: usize,
-    /// Maximum memory for workspace cache in bytes
-    pub max_bytes: usize,
-}
-
-impl Default for WorkspaceCacheConfig {
-    fn default() -> Self {
-        Self {
-            max_files: 1_000,
-            max_bytes: 20 * 1024 * 1024, // 20MB
-        }
-    }
-}
-
-/// Combined cache configuration for all workspace caches.
-#[derive(Clone, Debug, Default)]
-pub struct CombinedWorkspaceCacheConfig {
-    /// AST node cache configuration
-    pub ast: AstCacheConfig,
-    /// Symbol cache configuration
-    pub symbol: SymbolCacheConfig,
-    /// Workspace cache configuration
-    pub workspace: WorkspaceCacheConfig,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_cache_insert_get() {
-        let cache = BoundedLruCache::<String, String>::default();
-        cache.insert("key1".to_string(), "value1".to_string());
-        assert_eq!(cache.get(&"key1".to_string()), Some("value1".to_string()));
-    }
-
-    #[test]
-    fn test_cache_miss() {
-        let cache = BoundedLruCache::<String, String>::default();
-        assert_eq!(cache.get(&"nonexistent".to_string()), None);
-    }
-
-    #[test]
-    fn test_cache_eviction() {
-        let config = CacheConfig { max_items: 2, max_bytes: 100, ttl: None };
-        let cache = BoundedLruCache::<String, String>::new(config);
-
-        cache.insert("key1".to_string(), "value1".to_string());
-        cache.insert("key2".to_string(), "value2".to_string());
-        cache.insert("key3".to_string(), "value3".to_string());
-
-        // key1 should be evicted (LRU)
-        assert_eq!(cache.get(&"key1".to_string()), None);
-        assert_eq!(cache.get(&"key2".to_string()), Some("value2".to_string()));
-        assert_eq!(cache.get(&"key3".to_string()), Some("value3".to_string()));
-    }
-
-    #[test]
-    fn test_cache_stats() {
-        let cache = BoundedLruCache::<String, String>::default();
-        cache.insert("key1".to_string(), "value1".to_string());
-
-        cache.get(&"key1".to_string());
-        cache.get(&"key2".to_string());
-
-        let stats = cache.stats();
-        assert_eq!(stats.hits, 1);
-        assert_eq!(stats.misses, 1);
-        assert_eq!(stats.hit_rate, 0.5);
-    }
-
-    #[test]
-    fn test_cache_clear() {
-        let cache = BoundedLruCache::<String, String>::default();
-        cache.insert("key1".to_string(), "value1".to_string());
-        cache.clear();
-
-        assert!(cache.is_empty());
-        assert_eq!(cache.len(), 0);
-    }
-
-    #[test]
-    fn test_cache_remove() {
-        let cache = BoundedLruCache::<String, String>::default();
-        cache.insert("key1".to_string(), "value1".to_string());
-        assert_eq!(cache.remove(&"key1".to_string()), Some("value1".to_string()));
-        assert_eq!(cache.get(&"key1".to_string()), None);
-    }
-
-    #[test]
-    fn test_estimate_size_string() {
-        let s = "hello world";
-        assert_eq!(s.estimate_size(), 11);
-    }
-
-    #[test]
-    fn test_estimate_size_vec() {
-        let v = vec!["hello", "world"];
-        assert_eq!(v.estimate_size(), 10);
-    }
-}
+pub use perl_workspace_cache::*;
